# Simplify double-spending detection