--- a/WalletWasabi/Blockchain/TransactionProcessing/TransactionProcessor.cs
+++ b/WalletWasabi/Blockchain/TransactionProcessing/TransactionProcessor.cs
@@ -108,22 +108,10 @@ namespace WalletWasabi.Blockchain.TransactionProcessing
 					var doubleSpends = new List<SmartCoin>();
 					foreach (SmartCoin coin in Coins.AsAllCoinsView())
 					{
-						var spent = false;
-						foreach (TxoRef spentOutput in coin.SpentOutputs)
+						var spentOutputs = coin.SpentOutputs.Intersect(tx.Transaction.Inputs.ToTxoRefs());
+						if(spentOutputs.Any())
 						{
-							foreach (TxIn txIn in tx.Transaction.Inputs)
-							{
-								if (spentOutput.TransactionId == txIn.PrevOut.Hash && spentOutput.Index == txIn.PrevOut.N) // Do not do (spentOutput == txIn.PrevOut), it's faster this way, because it won't check for null.
-								{
-									doubleSpends.Add(coin);
-									spent = true;
-									break;
-								}
-							}
-							if (spent)
-							{
-								break;
-							}
+							doubleSpends.Add(coin);
 						}
 					}