Re-write hard-to-understand code.