@lontivero 2019-12-30
Sure. It would help a lot if someone could review and approve this PR: https://github.com/zkSNACKs/WalletWasabi/pull/2439