@lontivero 2019-12-30
Closing for performance reasons.

## Master branch

```
|  Method |     Job |     Toolchain | TRANSACTIONS |       Mean |     Error |    StdDev | Rank |
|-------- |-------- |-------------- |------------- |-----------:|----------:|----------:|-----:|
| Process | Default | netcoreapp3.0 |          100 |   1.631 ms | 0.0198 ms | 0.0175 ms |    1 |
| Process |    Core |       Default |          100 |   1.718 ms | 0.0328 ms | 0.0438 ms |    2 |
| Process | Default | netcoreapp3.0 |         1000 |  17.340 ms | 0.3432 ms | 0.4462 ms |    4 |
| Process |    Core |       Default |         1000 |  16.800 ms | 0.3315 ms | 0.3256 ms |    3 |
| Process | Default | netcoreapp3.0 |        10000 | 174.926 ms | 3.4253 ms | 5.8164 ms |    5 |
| Process |    Core |       Default |        10000 | 181.490 ms | 3.6190 ms | 5.8440 ms |    6 |
```

## PR

```
|  Method |     Job |     Toolchain | TRANSACTIONS |       Mean |      Error |     StdDev | Rank |
|-------- |-------- |-------------- |------------- |-----------:|-----------:|-----------:|-----:|
| Process | Default | netcoreapp3.0 |          100 |   4.822 ms |  0.0942 ms |  0.0967 ms |    1 |
| Process |    Core |       Default |          100 |   5.036 ms |  0.1005 ms |  0.1706 ms |    2 |
| Process | Default | netcoreapp3.0 |         1000 |  53.582 ms |  1.0588 ms |  1.6170 ms |    3 |
| Process |    Core |       Default |         1000 |  53.129 ms |  1.0141 ms |  0.9486 ms |    3 |
| Process | Default | netcoreapp3.0 |        10000 | 512.062 ms | 10.1924 ms | 16.1663 ms |    4 |
| Process |    Core |       Default |        10000 | 517.361 ms | 10.1793 ms | 16.4376 ms |    4 |
```